+++ Allegro_Spin/Angle.cpp
@@ -9,7 +9,7 @@ Angle::Angle():
 	inc(0),
 	defaultVal(0)
 {}
-Angle::Angle(double v0) :
+Angle::Angle(float v0) :
 	x(0),
 	y(0),
 	value(v0),
@@ -18,7 +18,7 @@ Angle::Angle(double v0) :
 	inc(0),
 	defaultVal(v0)
 {}
-Angle::Angle(double v0, double i0):
+Angle::Angle(float v0, float i0):
 	x(0),
 	y(0),
 	value(v0),
@@ -27,7 +27,7 @@ Angle::Angle(double v0, double i0):
 	inc(i0),
 	defaultVal(v0)
 {}
-Angle::Angle(double x0, double y0, double v0, double i0) :
+Angle::Angle(float x0, float y0, float v0, float i0) :
 	x(x0),
 	y(y0),
 	value(v0),
@@ -37,7 +37,7 @@ Angle::Angle(double x0, double y0, double v0, double i0) :
 	defaultVal(v0)
 {}
 
-void Angle::Inc(double side)//side == 1 or -1; -1 to left, 1 to right
+void Angle::Inc(float side)//side == 1 or -1; -1 to left, 1 to right
 {
 	if (value >= 360 + defaultVal) {
 		value -= 360;
@@ -49,19 +49,19 @@ void Angle::Inc(double side)//side == 1 or -1; -1 to left, 1 to right
 	cos = std::cos(value * pi / 180);
 	sin = std::sin(value * pi / 180);
 }
-double Angle::GetVal() const
+float Angle::GetVal() const
 {
 	return value;
 }
-double Angle::GetCos() const
+float Angle::GetCos() const
 {
 	return cos;
 }
-double Angle::GetSin() const
+float Angle::GetSin() const
 {
 	return sin;
 }
-void Angle::SetValue(double n){
+void Angle::SetValue(float n){
 	value = n;
 	cos = std::cos(value * pi / 180);
 	sin = std::sin(value * pi / 180);
