+++ Allegro_Spin/Poly3d.hpp
@@ -5,26 +5,26 @@
 class Poly3d
 {
 private:
-	double defaultHeight;
-	const double pi = 3.14159265;
+	float defaultHeight;
+	const float pi = 3.14159265;
 public:
-	double x;
-	double y;
-	double height;
-	double width;
+	float x;
+	float y;
+	float height;
+	float width;
 	EllipseAngles ellipse;
 	EllipseAngles ellipse2;
 	std::vector<Poly2d>Poly2ds;
-	Poly3d(double x0, double y0, double h0, double w0, double e1ry0, double e2rx0, double n);
-	Poly3d(double x0, double y0, double h0, double w0, double e1ry0, double e2rx0, double n, double e1i0, double e2i0);
-	void Scale(double n);
+	Poly3d(float x0, float y0, float h0, float w0, float e1ry0, float e2rx0, float n);
+	Poly3d(float x0, float y0, float h0, float w0, float e1ry0, float e2rx0, float n, float e1i0, float e2i0);
+	void Scale(float n);
 	/*
-	void SetX(double n);
-	void SetY(double n);
-	void IncX(double n);
-	void IncY(double n);
-	double GetX() const;
-	double GetY() const;
+	void SetX(float n);
+	void SetY(float n);
+	void IncX(float n);
+	void IncY(float n);
+	float GetX() const;
+	float GetY() const;
 	*/
 	void Move();
 	void Draw();
